Bottle Sprite and Character Controller
The sprite for the bottle character was completed with animations for moving left and right as well as for jumping up and down. Also the character controller was fixed so the character can now move left and right as well as jump. The fighting portion of this games still needs to be implemented in every way.